+++ src/Dereferencer.php
@@ -2,10 +2,6 @@
 
 namespace League\JsonGuard;
 
-use League\JsonGuard\Loaders\CurlWebLoader;
-use League\JsonGuard\Loaders\FileGetContentsWebLoader;
-use League\JsonGuard\Loaders\FileLoader;
-
 /**
  * The Dereferencer resolves all external $refs and replaces
  * internal references with Reference objects.
@@ -13,17 +9,18 @@ use League\JsonGuard\Loaders\FileLoader;
 class Dereferencer
 {
     /**
-     * @var array
+     * @var LoaderManager
      */
-    private $loaders;
+    private $loaderManager;
 
     /**
      * Create a new Dereferencer.
+     *
+     * @param LoaderManager $loaderManager
      */
-    public function __construct()
+    public function __construct(LoaderManager $loaderManager = null)
     {
-        $this->registerFileLoader();
-        $this->registerDefaultWebLoaders();
+        $this->loaderManager = $loaderManager ?: new LoaderManager();
     }
 
     /**
@@ -48,66 +45,11 @@ class Dereferencer
     }
 
     /**
-     * Register a Loader for the given prefix.
-     *
-     * @param Loader $loader
-     * @param string $prefix
+     * @return LoaderManager
      */
-    public function registerLoader(Loader $loader, $prefix)
+    public function getLoaderManager()
     {
-        $this->loaders[$prefix] = $loader;
-    }
-
-    /**
-     * Get all registered loaders, keyed by the prefix they are registered to load schemas for.
-     *
-     * @return Loader[]
-     */
-    public function getLoaders()
-    {
-        return $this->loaders;
-    }
-
-    /**
-     * Get the loader for the given prefix.
-     *
-     * @param string $prefix
-     *
-     * @return Loader
-     * @throws \InvalidArgumentException
-     */
-    public function getLoader($prefix)
-    {
-        if (!array_key_exists($prefix, $this->loaders)) {
-            throw new \InvalidArgumentException(sprintf('A loader is not registered for the prefix "%s"', $prefix));
-        }
-
-        return $this->loaders[$prefix];
-    }
-
-    /**
-     * Register the default file loader.
-     */
-    private function registerFileLoader()
-    {
-        $this->loaders['file'] = new FileLoader();
-    }
-
-    /**
-     * Register the default web loaders.  If the curl extension is loaded,
-     * the CurlWebLoader will be used.  Otherwise the FileGetContentsWebLoader
-     * will be used.  You can override this by registering your own loader
-     * for the 'http' and 'https' protocols.
-     */
-    private function registerDefaultWebLoaders()
-    {
-        if (function_exists('curl_init')) {
-            $this->loaders['https'] = new CurlWebLoader('https://');
-            $this->loaders['http']  = new CurlWebLoader('http://');
-        } else {
-            $this->loaders['https'] = new FileGetContentsWebLoader('https://');
-            $this->loaders['http']  = new FileGetContentsWebLoader('http://');
-        }
+        return $this->loaderManager;
     }
 
     /**
@@ -242,7 +184,7 @@ class Dereferencer
     private function loadExternalRef($reference)
     {
         list($prefix, $path) = parse_external_ref($reference);
-        $loader = $this->getLoader($prefix);
+        $loader = $this->loaderManager->getLoader($prefix);
         $schema = $loader->load($path);
 
         return $schema;